Deserialize NaiveDate directly from Strings
